requirements.txt
langchain
openai
python-dotenv
sentence_transformers
pypdf
faiss-cpu
tqdm
#openzeppelin-contracts
#faiss-gpu
#python-telegram-bot 
